Code Styling: Rename patch handlers
