+++ skytemple_files/patch/handler/complete_team_control.py
