+++ skytemple_files/patch/handler/partners_trigger_hidden_traps.py
