+++ skytemple_files/patch/handler/reduce_jumpcut_pause_time.py
